Update profile component to use model classes
