+++ src/features/organisation-profile/index.tsx
@@ -31,8 +31,8 @@ export default function OrgProfilePage() {
         <OrgBreadCrumb org={org} />
         <Summary />
         <Services />
-        <Profile org={org} />
-        {org !== undefined && <SimilarOrgs org={org} />}
+        {org !== undefined && <Profile org={org} />}
+        {/* {org !== undefined && <SimilarOrgs org={org} />} */}
         <Box className="maximise-width" minH="37.33vh">
           <Footer />
         </Box>
